requirements.txt
ultralytics~=8.3.28
Pillow~=10.3.0
numpy~=1.24.4
pandas~=2.0.3
matplotlib~=3.7.3
torch~=2.1.1
torchvision~=0.16.1